refactor: some api classes
