+++ sonar-api/src/main/java/xyz/jonesdev/sonar/api/Sonar.java
@@ -24,7 +24,6 @@ import xyz.jonesdev.sonar.api.fallback.Fallback;
 import xyz.jonesdev.sonar.api.logger.Logger;
 import xyz.jonesdev.sonar.api.server.ServerWrapper;
 import xyz.jonesdev.sonar.api.verbose.Verbose;
-import xyz.jonesdev.sonar.api.version.SonarVersion;
 
 import java.text.DecimalFormat;
